update mapper with additional fields

+++ src/main/groovy/edu/oregonstate/mist/personsapi/core/JobObject.groovy
@@ -26,7 +26,7 @@ class JobObject {
     BigDecimal hourlyRate
     BigDecimal hoursPerPay
     BigDecimal assignmentSalary
-    Integer paysPerYear
+    BigDecimal paysPerYear
     BigDecimal annualSalary
     List<LaborDistribution> laborDistribution
 }
